Add files via upload
This is a simple web app I made that takes a startNumber, endNumber, and a stepNumber and counts by the stepNumber through the start and endNumber. Error messages pop up if any input field is not a number, if endNumber is equal to or smaller than the startNumber or if anything is left blank.